+++ cloudsplaining/scan/inline_policy.py
@@ -26,8 +26,8 @@ class InlinePolicy:
             PolicyId=self.policy_id,
             PolicyDocument=self.policy_document.json,
             PrivilegeEscalation=self.policy_document.allows_privilege_escalation,
-            DataExfiltrationActions=self.policy_document.allows_data_leak_actions,
-            PermissionsManagementActions=self.policy_document.permissions_management_without_constraints,
+            DataExfiltration=self.policy_document.allows_data_exfiltration_actions,
+            ResourceExposure=self.policy_document.permissions_management_without_constraints,
         )
         return result
 
@@ -39,8 +39,8 @@ class InlinePolicy:
             PolicyId=self.policy_id,
             PolicyDocument=self.policy_document.json,
             PrivilegeEscalation=self.policy_document.allows_privilege_escalation,
-            DataExfiltrationActions=self.policy_document.allows_data_leak_actions,
-            PermissionsManagementActions=self.policy_document.permissions_management_without_constraints,
+            DataExfiltration=self.policy_document.allows_data_exfiltration_actions,
+            ResourceExposure=self.policy_document.permissions_management_without_constraints,
             InfrastructureModification=self.policy_document.all_allowed_unrestricted_actions
         )
         return result