+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 * Docker
 
-## 0.1.8 (Unreleased)
+## 0.1.8 (2020-08-27)
 * UI: The Exclusions configuration was not showing up in the report due to a typo
 * Changed `--input` flag to `--input-file` for all commands
 * Fixed bug in scan-policy-file command (#79)